hello_world/requirements.txt
requests
flask-lambda-python36
boto3
werkzeug==0.12.2